pyproject.toml
[project]
name = "symba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "James Ashford", email = "ashfordj@rocketmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.2",
    "flask>=3.1.3",
    "gitpython>=3.1.46",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pyinform>=0.2.0",
    "scipy>=1.17.1",
]

[project.scripts]
symba = "symba:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
